Ensure message block vector length is as expected
Message block vector length has to match block size (128) so that we
don't blow up when evaluating m vector later. In the original
implementation, F was an internal function and other public functions
calling it were making sure the correct vector is passed. Here, after
making F public, we need to carve this requirement in the interface.