t offset counter is no longer passed as a pointer
Since the offset counter is not internally modified there is not reason
for passing the array as a pointer. I have also renamed `c` field in
test vectors to `t`. `c` was used as offset counter name in the original
implementation while RFC uses `t`. Since we decided to stick to the RFC
naming, we should use it consistenly everywhere.